[#102]refactor: 필요없는 allobject 삭제
