+++ src/pages/Home/components/RightSection/services/addCodeFlow.ts
@@ -1,4 +1,4 @@
-import { AllObjectItem } from "@/pages/Home/types/allObjectItem";
+import { AllObjectItem } from "@/pages/Home/types/codeFlow/allObjectItem";
 
 // 새로운 객체를 CodeFlow에 추가하는 함수
 export const addCodeFlow = (
