+++ src/pages/Home/components/RightSection/services/insertEqualToDepth.ts
@@ -1,4 +1,4 @@
-import { AllObjectItem } from "@/pages/Home/types/allObjectItem";
+import { AllObjectItem } from "@/pages/Home/types/codeFlow/allObjectItem";
 
 export const insertEqualToDepth = (
   codeFlows: AllObjectItem[],
