+++ src/pages/Home/components/RightSection/services/refreshCodeFlow.ts
@@ -1,4 +1,4 @@
-import { AllObjectItem } from "@/pages/Home/types/allObjectItem";
+import { AllObjectItem } from "@/pages/Home/types/codeFlow/allObjectItem";
 
 export const refreshCodeFlow = (
   codeFlows: AllObjectItem[], //현제 코드흐름 시각화 정보를 담고 있는 리스트
