+++ src/pages/Home/components/RightSection/services/turnLight.ts
@@ -1,5 +1,5 @@
 import { ActivateItem } from "@/pages/Home/types/activateItem";
-import { AllObjectItem } from "@/pages/Home/types/allObjectItem";
+import { AllObjectItem } from "@/pages/Home/types/codeFlow/allObjectItem";
 
 // 하이라이트 효과를 줘야하는 부분을 표시해주는 함수
 export const turnLight = (
